requirements.txt
fastapi==0.128.0
uvicorn==0.39.0
sqlalchemy==2.0.46
pydantic==2.12.5
python-dotenv==1.2.1
python-multipart==0.0.20
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
pyotp==2.9.0
qrcode==8.2
pandas==2.3.3
numpy==2.0.2
yfinance==1.2.0
feedparser==6.0.12
requests==2.32.5
httpx==0.28.1
jinja2==3.1.6
email-validator==2.3.0
pytest==8.4.2
